pyproject.toml
[project]
name = "django-template"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "LucasGabrielAntonete", email = "lucasantonete@gmail.com"},
]
dependencies = [
    "django>=5.0.6",
    "djangorestframework>=3.15.1",
    "setuptools>=70.0.0",
    "djangorestframework-simplejwt>=5.3.1",
    "python-magic>=0.4.27",
    "Pillow>=10.3.0",
    "drf-spectacular>=0.27.2",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.1",
    "django-cors-headers>=4.3.1",
    "netifaces>=0.11.0",
    "django-filter>=24.2",
    "pytz>=2024.1",
    "django-extensions>=3.2.3",
    "pygraphviz>=1.13",
    "pyparsing>=3.1.2",
    "pydot>=2.0.0",
    "django-simple-history>=3.7.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
